Moved the PolygonPlotter in hazardlib
